+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Documentation - Add guides on generic hardware backend and developing a new hardware backend in quantify (!817)
 - Qblox backend - Move `_insert_update_parameters` from `QbloxBaseModule` to `Sequencer`. (!912)
 - Serialization - Allow `"deserialization_type"` in the serialized dictionary to be the fully qualified name of the object, such that it can be imported via `import_python_object_from_string`. (!905)
+- Compilation - Allow gates to be device compiled to schedules. (!904)
+- Operations - Add `H` Hadamard gate to gate library. (!904)
 
 ## 0.18.1 (2024-02-22)